third_party/WebKit/LayoutTests/batterystatus/window-property-expected.txt

Tests that the window.navigator.battery properties are present.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL typeof navigator.webkitBattery == 'object' should be true. Was false.
FAIL hasOnBatteryProperty() should be true. Was false.
FAIL navigator.webkitBattery.hasOwnProperty('onchargingchange') should be true. Threw exception TypeError: Cannot call method 'hasOwnProperty' of undefined
FAIL navigator.webkitBattery.hasOwnProperty('onchargingtimechange') should be true. Threw exception TypeError: Cannot call method 'hasOwnProperty' of undefined
FAIL navigator.webkitBattery.hasOwnProperty('ondischargingtimechange') should be true. Threw exception TypeError: Cannot call method 'hasOwnProperty' of undefined
FAIL navigator.webkitBattery.hasOwnProperty('onlevelchange') should be true. Threw exception TypeError: Cannot call method 'hasOwnProperty' of undefined
PASS successfullyParsed is true

TEST COMPLETE
